requirements.txt
azure-cosmos
python-dotenv
fastapi
uvicorn
pytest
azure-identity
azure-keyvault-secrets
black
isort